+++ galacteek/core/ps.py
@@ -78,9 +78,7 @@ mSubscriber = psSubscriber('main')
 
 
 def hubPublish(key, message):
-    log.debug(f'hubPublish ({key}) : {message}')
-
-    # Hub publishing should now always happen in the main loop
+    # Publishing to the hub should now always happen in the main loop
     app = runningApp()
     app.loop.call_soon_threadsafe(gHub.publish, key, message)