+++ galacteek/dweb/channels/ontolochain.py
@@ -13,7 +13,6 @@ from galacteek.core import runningApp
 from galacteek.core import utcDatetimeIso
 
 from galacteek.ld import gLdBaseUri
-from galacteek.ld.signatures import jsonldsig
 from galacteek.ld import ontolochain
 
 from . import GOntoloObject
@@ -69,10 +68,8 @@ class OntoloChainInterface(object):
 
             dag['verificationMethod'] = f'{ipid.did}#keys-1'
 
-            signed = jsonldsig.sign(dag, privKey.exportKey())
-
             # Store in IPFS
-            cid = await ipfsop.dagPut(signed)
+            cid = await ipfsop.dagPut(dag)
 
             # Notify the RDF service
             await runningApp().s.rdfStore(