+++ galacteek/ld/rdf/terms/DID.py
@@ -1,0 +1,3 @@
+from rdflib import URIRef
+
+documentIpfsCid = URIRef('https://w3id.org/did#documentIpfsCid')